+++ src/error.rs
@@ -22,6 +22,9 @@ pub enum RpxyError {
   #[error("TCP/UDP Proxy Layer Error: {0}")]
   Proxy(String),
 
+  #[error("LoadBalance Layer Error: {0}")]
+  LoadBalance(String),
+
   #[error("I/O Error")]
   Io(#[from] io::Error),